[REFACTOR]

+++ deploy.sh
@@ -10,8 +10,8 @@ CONTAINER_NAME="be-auth-user-staff"
 deploy_version() {
   echo "Deploying version $VERSION..."
 
-  # Check if a container with the same name is already running
-  if [ "$(sudo docker ps -q -f name=$CONTAINER_NAME)" ]; then
+  # Check if a container with the same name is already running or exists
+  if [ "$(sudo docker ps -a -q -f name=$CONTAINER_NAME)" ]; then
     echo "Stopping and removing existing container..."
     sudo docker stop $CONTAINER_NAME
     sudo docker rm $CONTAINER_NAME
